# openapi-spec.yaml
openapi: 3.0.3
info:
  title: MyLab Platform API
  description: Complete API specification for the MyLab Platform backend
  version: 1.1.0
  contact:
    name: MyLab Platform Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.mylab-platform.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /analysis-requests:
    post:
      summary: Create analysis request
      description: Create new analysis request for external lab
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequestRequest'
      responses:
        '201':
          description: Analysis request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AnalysisRequest'
  /analysis-requests/incoming:
    get:
      summary: Get incoming requests
      description: Get analysis requests TO current organization
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, rejected]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        '200':
          description: Incoming requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalysisRequestWithDetails'
  /analysis-requests/outgoing:
    get:
      summary: Get outgoing requests
      description: Get analysis requests FROM current organization
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, rejected]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        '200':
          description: Outgoing requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalysisRequestWithDetails'
  /analysis-requests/{id}:
    get:
      summary: Get analysis request
      description: Get analysis request details
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis request details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AnalysisRequestWithDetails'

    patch:
      summary: Update request status
      description: Update analysis request status
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [in_progress, completed]
                notes:
                  type: string
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AnalysisRequest'
  /analysis-requests/{id}/accept:
    post:
      summary: Accept request
      description: Accept analysis request and assign
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assignedTo
              properties:
                assignedTo:
                  type: string
                  format: uuid
                  description: User ID to assign the request to
      responses:
        '200':
          description: Request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AnalysisRequest'
  /analysis-requests/{id}/reject:
    post:
      summary: Reject request
      description: Reject analysis request
      tags: [Analysis Requests]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: Reason for rejection
      responses:
        '200':
          description: Request rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AnalysisRequest'

  # ============================================================================
  # FILES
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalysisType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Analysis type name
        description:
          type: string
          description: Detailed description
        methodology:
          type: string
          description: Analysis methodology
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AnalysisRequestRequest:
      type: object
      required:
        - toOrganizationId
        - sampleId
        - analysisTypeId
        - description
      properties:
        toOrganizationId:
          type: string
          format: uuid
          description: ID of organization to send request to
        sampleId:
          type: string
          format: uuid
          description: ID of sample for analysis
        analysisTypeId:
          type: string
          format: uuid
          description: ID of requested analysis type
        description:
          type: string
          description: Detailed description of request
        methodologyRequirements:
          type: string
          description: Specific methodology requirements
        parameters:
          type: object
          description: Additional analysis parameters
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        dueDate:
          type: string
          format: date
          description: Requested completion date
        estimatedDuration:
          type: string
          description: Estimated analysis duration
        notes:
          type: string
          description: Additional notes
    AnalysisRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        fromOrganizationId:
          type: string
          format: uuid
        toOrganizationId:
          type: string
          format: uuid
        sampleId:
          type: string
          format: uuid
        analysisTypeId:
          type: string
          format: uuid
        description:
          type: string
        methodologyRequirements:
          type: string
        parameters:
          type: object
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, rejected]
        dueDate:
          type: string
          format: date
        estimatedDuration:
          type: string
        assignedTo:
          type: string
          format: uuid
        notes:
          type: string
        createdBy:
          type: string
          format: uuid
        acceptedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AnalysisRequestWithDetails:
      allOf:
        - $ref: '#/components/schemas/AnalysisRequest'
        - type: object
          properties:
            sample:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
            analysisType:
              $ref: '#/components/schemas/AnalysisType'
            fromOrganization:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
            toOrganization:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
            assignedToUser:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                email:
                  type: string

    # ============================================================================
    # FILE SCHEMAS
    # ============================================================================
...